feat(harnesses): host the model-credential ladder
The ladder moves out of the umbrella so the standalone agent runtime can
fill its model slot from VENDO_API_KEY without depending on
`@vendoai/vendo`. It is a move, not a rewrite: the two files land under
`src/inference/` and their behaviour is unchanged.

- `@vendoai/harnesses/inference` is the ladder (`vendoModel`), and
  `@vendoai/harnesses/inference/credential` its pure half. Two subpaths
  because they have different portability: the ladder resolves the host's
  provider install from disk, while the credential rungs are pure and the
  boot summary reads them from a graph that must bundle for a Worker.
- vendo keeps both file names as doors onto the new home. It has to: the
  `#dev-creds/model` condition names `dev-creds/model.js` for Node and the
  edge twin for web-standard runtimes, and that condition is what keeps
  Node resolution out of a Worker bundle. The portability gate confirms
  the worker graph is still Node-leg-free.
- The Cloud rung reads `VENDO_CLOUD_URL` itself instead of calling the
  umbrella's `resolveCloudBaseUrl` — same default, same trailing-slash
  trim; the edge back to `@vendoai/vendo` would be a cycle.
- `@vendoai/apps` joins the harnesses layer for the one rule the ladder
  cannot restate without owning a second copy of it: which model ids
  accept sampling parameters, re-decided against the resolved rung. apps
  depends on core alone, so there is no cycle.
- `@ai-sdk/anthropic` becomes a real dependency: the provider fallback
  resolves from this package's own module context now, which is what
  lights up live chat with nothing installed in the host repo.
